refactor: extract renderer facade state carriers
